megaco/src/flex/Makefile.in: Support parallel make
Make sure that directories are created before they are used.

I tried to use order-only prerequisites to create the directories,
but run into two problems: First, order-only prerequisites are
only implemented in Make 3.80 and later. Second, on a computer
running Solaris/Intel 2.8 (with Make 3.80), order-only prerequisites
seemed to work like ordinary prerequisites, causing targets to
be re-built if the timestamp for the directory changed.

Therefore, using a shell command to run mkdir seems to be the
more portable solution.